basic acq fnc implemented; active learning placed in separate file
