+++ docs/refs.bib
@@ -223,14 +223,46 @@
 }
 
 @InProceedings{kandasamy18a,
-  title = {Parallelised Bayesian Optimisation via Thompson Sampling},
-  author = {Kandasamy, Kirthevasan and Krishnamurthy, Akshay and Schneider, Jeff and Poczos, Barnabas},
-  booktitle = {Proceedings of the Twenty-First International Conference on Artificial Intelligence and Statistics},
-  pages = {133--142},
-  year = {2018},
-  editor = {Storkey, Amos and Perez-Cruz, Fernando},
-  volume = {84},
-  series = {Proceedings of Machine Learning Research},
-  publisher = {PMLR},
-  url = {https://proceedings.mlr.press/v84/kandasamy18a.html}
+    title = {Parallelised Bayesian Optimisation via Thompson Sampling},
+    author = {Kandasamy, Kirthevasan and Krishnamurthy, Akshay and Schneider, Jeff and Poczos, Barnabas},
+    booktitle = {Proceedings of the Twenty-First International Conference on Artificial Intelligence and Statistics},
+    pages = {133--142},
+    year = {2018},
+    editor = {Storkey, Amos and Perez-Cruz, Fernando},
+    volume = {84},
+    series = {Proceedings of Machine Learning Research},
+    publisher = {PMLR},
+    url = {https://proceedings.mlr.press/v84/kandasamy18a.html}
+}
+
+@article{ranjan2008sequential,
+    title={Sequential experiment design for contour estimation from complex computer codes},
+    author={Ranjan, Pritam and Bingham, Derek and Michailidis, George},
+    journal={Technometrics},
+    volume={50},
+    number={4},
+    pages={527--541},
+    year={2008},
+    publisher={Taylor \& Francis}
+}
+
+@article{bichon2008efficient,
+    title={Efficient global reliability analysis for nonlinear implicit performance functions},
+    author={Bichon, Barron J and Eldred, Michael S and Swiler, Laura Painton and Mahadevan, Sandaran and McFarland, John M},
+    journal={AIAA journal},
+    volume={46},
+    number={10},
+    pages={2459--2468},
+    year={2008}
+}
+
+@article{bect2012sequential,
+    title={Sequential design of computer experiments for the estimation of a probability of failure},
+    author={Bect, Julien and Ginsbourger, David and Li, Ling and Picheny, Victor and Vazquez, Emmanuel},
+    journal={Statistics and Computing},
+    volume={22},
+    number={3},
+    pages={773--793},
+    year={2012},
+    publisher={Springer}
 }
